clean up archive files
- when archive has multi parts, dont try
-clean up code standars

+++ main.php
@@ -4,12 +4,12 @@ class MainProgram
 	public function run($config) {
 
 		$lineSeperator = "<br>";
-		if (php_sapi_name() == 'cli'){
+		if (php_sapi_name() == 'cli') {
 		  $lineSeperator = "\n\r";
 		}
 
 		echo "Script ran at: " . date('Y-m-d H:i:s') . $lineSeperator;
-		if (file_exists($config->lockFile)){
+		if (file_exists($config->lockFile)) {
 		 exit('Lock File Exists');
 		}
 
@@ -23,11 +23,11 @@ class MainProgram
 
 		$recurseDir = new RecursiveDirectoryIterator($config->scanningDir);
 		$objects = new RecursiveIteratorIterator($recurseDir, RecursiveIteratorIterator::SELF_FIRST);
-		foreach($objects as $name => $object){
+		foreach($objects as $name => $object) {
 			$pathinfo = pathinfo($name);
 
 			$extension = '';
-			if (isset($pathinfo['extension'])){
+			if (isset($pathinfo['extension'])) {
 				$extension = strtolower($pathinfo['extension']);
 			}
 
@@ -45,7 +45,7 @@ class MainProgram
 				}
 			}
 		}
-		foreach ($videos as $video){
+		foreach ($videos as $video) {
 		
 		  //make the directory
 		  if ($video->isSeries()) {
@@ -53,7 +53,7 @@ class MainProgram
 
 			$videoLocation = $directory . $video->getSeries() ." - s" .$video->getSeason() ."e". $video->getEpisode() . " (" .$video->getVideoQuality() . "p).".$video->getFileExtension();
 
-			if ($config->debug){
+			if ($config->debug) {
 			  echo "Series: ". $video->getSeries() . $lineSeperator;
 			  echo "Season: ".$video->getSeason() . $lineSeperator;
 			  echo "Episode: ". $video->getEpisode() . $lineSeperator;
@@ -73,7 +73,7 @@ class MainProgram
 
 		  if ($config->prod && !file_exists($config->lockFile)){
 			//if it already exists, just delete it
-			if (!file_exists($videoLocation)){
+			if (!file_exists($videoLocation)) {
 			  @mkdir($directory, $permissions, true);
 			    rename($video->getFileLocation().$video->getFilename(), $videoLocation);
 			} else {
@@ -84,7 +84,7 @@ class MainProgram
 		  }
 		}
 
-		foreach ($audios as $audio){
+		foreach ($audios as $audio) {
 
 		  $artist = $audio->getArtist();
 		  $artist = (!empty($artist)) ? $artist . "/" : "various-artist/";
@@ -98,15 +98,15 @@ class MainProgram
 		  $directory = $config->musicDir . $artist . $album;
 
 		  //make sure a title exists, otherwise will have no idea what it is
-		  if ($audio->getTitle()  !== ""){
+		  if ($audio->getTitle()  !== "") {
 			$filename = $trackNumber . $audio->getTitle() . "." . $audio->getExtension();
 		  } else {
 			$filename = $audio->getFileName();
 		  }
 		  $newLocation = $directory . $filename;
 
-		  if ($config->prod && !file_exists($config->lockFile)){
-			if (!file_exists($newLocation)){
+		  if ($config->prod && !file_exists($config->lockFile)) {
+			if (!file_exists($newLocation)) {
 			  @mkdir($directory, $permissions, true);
 			    rename($audio->getFileLocation().$audio->getFileName(), $newLocation);
 			} else {
@@ -116,7 +116,7 @@ class MainProgram
 			$lockFlag = true;
 		  }
 
-		  if ($config->debug){
+		  if ($config->debug) {
 		  echo "Old file location: " . $audio->getFileLocation(). $audio->getFileName() . $lineSeperator;
 		  echo "New file location: " . $newLocation . $lineSeperator . $lineSeperator;
 		  }
@@ -128,26 +128,24 @@ class MainProgram
 			
 		  //move them into an new "Other to sort" folder
 		  $pathinfo = pathinfo($other);
-			if ($pathinfo['extension'] === 'rar'){
+			if ($pathinfo['extension'] === 'rar') {
 				$rarFile = rar_open($other);
 				$list = rar_list($rarFile);
 				foreach($list as $file) {
 					$entry = rar_entry_get($rarFile, $file->getName());
 					$entry->extract($config->scanningDir); // extract to the current dir
 					$archiveFiles[$pathinfo['filename']]['extracted'] = true;
-					//@todo - make sure the same 'files' aren't added twice (otherwise there will be a notice)
 					$archiveFiles[$pathinfo['filename']]['files'][] = $other; 
+					if ($config->debug){
+						echo "UnRared File: " . $other . $lineSeperator;
+					}
 				}
-
+				rar_close($rarFile);
 				//if the rar file still exists after extraction, remove it later - multi part rar files fall into this
 				if (file_exists($other)){
 					$archiveFiles[$pathinfo['filename']]['extracted'] = true;	
 					$archiveFiles[$pathinfo['filename']]['files'][] = $other;
 				}
-				if ($config->debug){
-					echo "UnRared File: " . $other . $lineSeperator;
-				}
-				rar_close($rarFile);
 			} else if ( 1 === preg_match ('/r[0-9]{2}/', $pathinfo['extension'])){
 			//get the filename and save it in an array. Once we unzip, we can delete it
 				$archiveFiles[$pathinfo['filename']]['files'][] = $other; 
@@ -164,9 +162,9 @@ class MainProgram
 					echo "To : " . $config->manualSort . $pathinfo['filename'] . "." . $pathinfo['extension'] . $lineSeperator . $lineSeperator;
 				}
 
-				if ($config->prod && !file_exists($config->lockFile)){
+				if ($config->prod && !file_exists($config->lockFile)) {
 					//if its a sample or the extension is ignore, just remove it
-					if (!in_array($pathinfo['extension'], $config->ignoreFiles) && strpos($pathinfo['filename'], 'sample') === false){
+					if (!in_array($pathinfo['extension'], $config->ignoreFiles) && strpos($pathinfo['filename'], 'sample') === false) {
 						rename($other, $config->manualSort . $pathinfo['filename'] . "." . $pathinfo['extension']);
 					} else {
 						unlink($other);
@@ -176,18 +174,19 @@ class MainProgram
 				}
 			}
 		}
-		
 		//====Cleanup====//
-
 		//Cleanup archive files if they've been extracted
-		foreach ($archiveFiles as $archive){
+		foreach ($archiveFiles as $archive) {
 			if ($archive['extracted'] === true) {
-				if (is_array($archive['files'])){
+				$archive['files'] = array_unique($archive['files']);
+				if (is_array($archive['files'])) {
 					foreach ($archive['files'] as $file){
-						if ($config->debug){
+						if ($config->debug) {
 							echo "Delete archive File: " . $file . $lineSeperator;
 						}
-					unlink($file);
+						if (file_exists($file)) {
+							unlink($file);
+						}
 					}
 				}
 			}
